refactor(sources): extract shared host address safety checks into Source CDK (#956)
Move the loopback/private/link-local and numeric-IPv4-literal host classification helpers out of the cosmo source into a shared internal/sourcecdk/host_safety.go using typed string/net.IP signatures. No emitted events, attributes, cursors, or runtime behavior change; the source keeps its own base_url policy and error messages. Lowers the cosmo no-growth ratchet 1104 -> 1012.
